moved files and removed unnecessary folders with single files;
