--- src/components/dashboard/StudentAnnouncements.js
@@ -1,9 +1,0 @@
-// import React from 'React'
-
-// export default StudentAnnouncement() {
-//     return(
-//         <div>
-//             <h3>be cool </h3>
-//         </div>
-//     )
-// }
