+++ src/components/stats/MentorStats.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
-import { BACKEND_URL, STATS_API } from "../../../constants/constants";
-import Footer from "../../Footer";
-import Navbar from "../../Navbar";
+import { BACKEND_URL, STATS_API } from "../../constants";
+import Footer from "../Footer";
+import Navbar from "../Navbar";
 
 function trim_message(message) {
   if (message)
