+++ src/components/tables/StudentsTable.js
@@ -2,10 +2,10 @@ import axios from "axios";
 import Fuse from "fuse.js";
 import React, { useEffect, useMemo, useState } from "react";
 import { useSortBy, useTable } from "react-table";
-import { STATS_API } from "../../../constants/constants";
-import { shuffleArray } from "../../../utils/shuffle";
-import Footer from "../../Footer.js";
-import Navbar from "../../Navbar.js";
+import { STATS_API } from "../../constants";
+import { shuffleArray } from "../../utils/shuffle";
+import Footer from "../Footer.js";
+import Navbar from "../Navbar.js";
 
 const debouncer = function (fn, delay) {
   let timer;
